Move return_non_coro_awaitable to conftest
